chromium/third_party/blink/web_tests/http/tests/images/document-policy-lossless-images-max-bpp-expected.txt

CONSOLE ERROR: Document policy violation: lossless-images-max-bpp is not allowed in this document.
